apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv>=1.0.0
openai>=1.0.0
